+++ pkg/kv/kvserver/store.go
@@ -2698,76 +2698,6 @@ func ReadMaxHLCUpperBound(ctx context.Context, engines []storage.Engine) (int64,
 	return hlcUpperBound, nil
 }
 
-// checkCanInitializeEngine ensures that the engine is empty except for a
-// cluster version, which must be present.
-func checkCanInitializeEngine(ctx context.Context, eng storage.Engine) error {
-	// See if this is an already-bootstrapped store.
-	ident, err := kvstorage.ReadStoreIdent(ctx, eng)
-	if err == nil {
-		return errors.Errorf("engine already initialized as %s", ident.String())
-	} else if !errors.HasType(err, (*kvstorage.NotBootstrappedError)(nil)) {
-		return errors.Wrap(err, "unable to read store ident")
-	}
-	// Engine is not bootstrapped yet (i.e. no StoreIdent). Does it contain a
-	// cluster version, cached settings and nothing else? Note that there is one
-	// cluster version key and many cached settings key, and the cluster version
-	// key precedes the cached settings.
-	//
-	// We use an EngineIterator to ensure that there are no keys that cannot be
-	// parsed as MVCCKeys (e.g. lock table keys) in the engine.
-	iter := eng.NewEngineIterator(storage.IterOptions{
-		KeyTypes:   storage.IterKeyTypePointsAndRanges,
-		UpperBound: roachpb.KeyMax,
-	})
-	defer iter.Close()
-	valid, err := iter.SeekEngineKeyGE(storage.EngineKey{Key: roachpb.KeyMin})
-	if !valid {
-		if err == nil {
-			return errors.New("no cluster version found on uninitialized engine")
-		}
-		return err
-	}
-	getMVCCKey := func() (storage.MVCCKey, error) {
-		if _, hasRange := iter.HasPointAndRange(); hasRange {
-			bounds, err := iter.EngineRangeBounds()
-			if err != nil {
-				return storage.MVCCKey{}, err
-			}
-			return storage.MVCCKey{}, errors.Errorf("found mvcc range key: %s", bounds)
-		}
-		var k storage.EngineKey
-		k, err = iter.EngineKey()
-		if err != nil {
-			return storage.MVCCKey{}, err
-		}
-		if !k.IsMVCCKey() {
-			return storage.MVCCKey{}, errors.Errorf("found non-mvcc key: %s", k)
-		}
-		return k.ToMVCCKey()
-	}
-	var k storage.MVCCKey
-	if k, err = getMVCCKey(); err != nil {
-		return err
-	}
-	if !k.Key.Equal(keys.StoreClusterVersionKey()) {
-		return errors.New("no cluster version found on uninitialized engine")
-	}
-	valid, err = iter.NextEngineKey()
-	for valid {
-		// Only allowed to find cached cluster settings on an uninitialized
-		// engine.
-		if k, err = getMVCCKey(); err != nil {
-			return err
-		}
-		if _, err := keys.DecodeStoreCachedSettingsKey(k.Key); err != nil {
-			return errors.Errorf("engine cannot be bootstrapped, contains key:\n%s", k.String())
-		}
-		// There may be more cached cluster settings, so continue iterating.
-		valid, err = iter.NextEngineKey()
-	}
-	return err
-}
-
 // GetReplica fetches a replica by Range ID. Returns an error if no replica is found.
 //
 // See also GetReplicaIfExists for a more perfomant version.
